+++ sys/src/9/port/lib.h
@@ -122,7 +122,7 @@ extern	char*	fmtstrflush(Fmt*);
  * one-of-a-kind
  */
 extern	char*	cleanname(char*);
-extern	ulong	getcallerpc(void*);
+extern	uintptr	getcallerpc(void*);
 
 extern	long	strtol(char*, char**, int);
 extern	ulong	strtoul(char*, char**, int);
